add primitive point deletion adjust Primitive::draw for primitives consisting of single point incinerate double <-> float warnings!

+++ lab05/headers/geometry/primitives/bezier_curve.h
@@ -5,7 +5,7 @@
 
 class BezierCurve : public Primitive
 {
-	float _step = 0.02;
+	float _step = (float)0.02;
 	void draw_skeleton(ImDrawList* draw_list, const ImVec2& offset);
 	ImVec2& nextBt3(const std::deque<ImVec2>& points, const float& t);
 	void draw_curve3(ImDrawList* draw_list, const std::deque<ImVec2>& points, const ImVec2& offset);
